feat: user authorization for creating and deleting user reviews. Admin has full authorization.

+++ app/controllers/user_reviews_controller.rb
@@ -1,9 +1,10 @@
 class UserReviewsController < ApplicationController
-  # after_action :verify_authorized, unless: :skip_pundit?
+  after_action :verify_authorized, unless: :skip_pundit?
 
   def new
     @cocktail = Cocktail.find(params[:cocktail_id])
     @user_review = UserReview.new
+    authorize @user_review
   end
 
   def create
@@ -11,7 +12,8 @@ class UserReviewsController < ApplicationController
     @user_review = UserReview.new(user_review_params)
     @user_review.cocktail = @cocktail
     @user_review.user = current_user
-    # raise
+    authorize @user_review
+
     if @user_review.save
       redirect_to cocktail_path(@cocktail)
     else
@@ -21,16 +23,18 @@ class UserReviewsController < ApplicationController
 
   def edit
     @user_review = UserReview.find(params[:id])
+    authorize @user_review
   end
 
   def update
     @user_review = UserReview.find(params[:id])
     @cocktail = @user_review.cocktail
+    authorize @user_review
 
     if @user_review.update(user_review_params)
       redirect_to cocktail_path(@cocktail)
     else
-      render edit, content: :unprocessable_content
+      render :edit, content: :unprocessable_content
     end
 
   end
@@ -39,6 +43,7 @@ class UserReviewsController < ApplicationController
   def destroy
     @user_review = UserReview.find(params[:id])
     @cocktail = @user_review.cocktail
+    authorize @user_review
 
     if @user_review.destroy
       redirect_to cocktail_path(@cocktail)
